Added regression tests for dlopen source
